app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
structlog==24.1.0
redis==5.0.8
qdrant-client==1.9.1
httpx==0.27.2
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
tiktoken==0.7.0
numpy==2.1.1
openai==1.45.0
azure-search-documents==11.6.0
azure-core==1.30.2
azure-identity==1.17.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.9

